+++ src/Exception/Client401UnauthorizedException.php
@@ -8,9 +8,9 @@ class Client401UnauthorizedException extends ProblemJsonException
 {
     public function __construct(
         string $type,
-        string $title = 'Request does not contain valid token, or anonymous user is disabled.',
+        string $title = 'Unauthorized',
         int $status = 401,
-        string $detail = '',
+        string $detail = 'Authorization for the request failed due to possible problems with the token (incorrect or expired), password (incorrect or changed), the user\'s unique identifier, or the user\'s status (e.g., missing, blocked, or deleted).',
         string $instance = null,
         Throwable $previous = null
     ) {