+++ src/Kernel.php
@@ -2,17 +2,10 @@
 
 namespace App;
 
-use App\DependencyInjection\TraceableEventDispatcherDependencyInjection;
 use Symfony\Bundle\FrameworkBundle\Kernel\MicroKernelTrait;
-use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\HttpKernel\Kernel as BaseKernel;
 
 class Kernel extends BaseKernel
 {
     use MicroKernelTrait;
-
-    protected function build(ContainerBuilder $container): void
-    {
-        $container->addCompilerPass(new TraceableEventDispatcherDependencyInjection());
-    }
 }