+++ tests/UnitTests/Factory/Exception/Client401UnauthorizedExceptionFactoryTest.php
@@ -1,0 +1,30 @@
+<?php
+
+namespace App\Tests\UnitTests\Factory\Exception;
+
+use App\Factory\Exception\Client401UnauthorizedExceptionFactory;
+use PHPUnit\Framework\TestCase;
+use Prophecy\Argument;
+use Prophecy\PhpUnit\ProphecyTrait;
+use Symfony\Component\Routing\Generator\UrlGeneratorInterface;
+
+class Client401UnauthorizedExceptionFactoryTest extends TestCase
+{
+    use ProphecyTrait;
+
+    public function testCreateFromTemplate(): void
+    {
+        $urlGenerator = $this->prophesize(UrlGeneratorInterface::class);
+        $urlGenerator->generate(Argument::cetera())->shouldBeCalledOnce()->willReturn('https://mock.dev/123');
+        $factory = new Client401UnauthorizedExceptionFactory($urlGenerator->reveal());
+
+        $exception = $factory->createFromTemplate();
+
+        $this->assertSame(401, $exception->getStatus());
+        $this->assertSame('Unauthorized', $exception->getTitle());
+        $this->assertSame('https://mock.dev/123', $exception->getType());
+        $this->assertSame("Authorization for the request failed due to possible problems with the token (incorrect or expired), password (incorrect or changed), the user's unique identifier, or the user's status (e.g., missing, blocked, or deleted).", $exception->getDetail());
+        $this->assertSame(null, $exception->getInstance());
+        $this->assertSame('', $exception->getMessage());
+    }
+}