+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- Add diverse unit tests.
+- Add timeout to some GitHub Action jobs.
 
 ## 0.0.32 - 2023-09-30
 ### Added